restructure src folder and script to main
