--- src/runOnce.ts
@@ -1,5 +1,0 @@
-import { prontoSellMagCommentScript } from "./script.js";
-
-(async () => {
-  await prontoSellMagCommentScript();
-})();
