+++ src/scripts/runOnce.ts
@@ -1,0 +1,5 @@
+import { prontoSellMagCommentScript } from "../functions/main.js";
+
+(async () => {
+  await prontoSellMagCommentScript();
+})();
